fix: resolve all 154 flutter analyze issues (0 errors, 0 warnings, 0 info)
Errors fixed (30):
- auto_sync_engine_monitor_screen: import AutoSyncEngineState, ConflictResolutionStrategy
- google_drive_logs_screen: remove extra closing brace breaking class structure
- settings_maintenance: import BookingDerivedFieldsService
- sync_health_screen: import SyncHealthReport, SyncHealthStatus, SyncHealthMonitor
- unified_sync_orchestrator: fix Future.wait type casting (11 map_value_type errors)
- google_drive_conflict_resolver: cast uuid to String

Warnings fixed (10):
- Remove dead_null_aware expressions (expenses_list, bookings_repository, payload_mapper, appwrite_sync_manager)
- Remove unused imports (secondary_backup_service, secondary_sync_tracker, smart_conflict_resolver)
- Remove unused element (_buildFallbackMessage)
- Remove duplicate import (settings_screen)
- Remove unnecessary cast (finance_screen)
- Comment out unrecognized diagnostic code (analysis_options.yaml)

Info fixed (114):
- Sort import directives alphabetically in 14 files
- Add const constructors, prefer const literals
- Remove redundant argument values matching defaults
- Fix eol_at_end_of_file in 3 files
- Replace final with const for constants
- Remove noop .toDouble() on double values
- Wrap unawaited futures with unawaited()
- Change void async return types to Future<void>
- Add library directive for dangling doc comment
- Replace deprecated cacheExtent usage
- Remove unnecessary braces in string interpolation
- Sort pubspec dependencies alphabetically
- Move constructors before non-constructor declarations

+++ mobile/lib/services/repositories/salary_withdrawals_repository.dart
@@ -73,12 +73,12 @@ class SalaryWithdrawalsRepository {
     // إشعارات فورية (fire-and-forget)
     unawaited(WhatsAppNotificationService.instance.notifyNewExpense(
       category: 'سحب راتب',
-      amount: amount.toDouble(),
+      amount: amount,
       description: reason,
     ));
     unawaited(TelegramNotificationService.instance.notifyNewExpense(
       category: 'سحب راتب',
-      amount: amount.toDouble(),
+      amount: amount,
       description: reason,
     ));
 
@@ -249,12 +249,12 @@ class SalaryWithdrawalsRepository {
         // إشعارات فورية (fire-and-forget) عند التحديث
         unawaited(WhatsAppNotificationService.instance.notifyNewExpense(
           category: 'سحب راتب',
-          amount: amount.toDouble(),
+          amount: amount,
           description: note ?? reasonText,
         ));
         unawaited(TelegramNotificationService.instance.notifyNewExpense(
           category: 'سحب راتب',
-          amount: amount.toDouble(),
+          amount: amount,
           description: note ?? reasonText,
         ));
       } else {
@@ -287,12 +287,12 @@ class SalaryWithdrawalsRepository {
         await _setExpenseIdRaw(newId, expenseId);
         unawaited(WhatsAppNotificationService.instance.notifyNewExpense(
           category: 'سحب راتب',
-          amount: amount.toDouble(),
+          amount: amount,
           description: note,
         ));
         unawaited(TelegramNotificationService.instance.notifyNewExpense(
           category: 'سحب راتب',
-          amount: amount.toDouble(),
+          amount: amount,
           description: note,
         ));
